Struct google_civicinfo2::AdministrativeBody

pub struct AdministrativeBody {
    pub absentee_voting_info_url: Option<String>,
    pub voting_location_finder_url: Option<String>,
    pub name: Option<String>,
    pub election_registration_confirmation_url: Option<String>,
    pub election_info_url: Option<String>,
    pub election_registration_url: Option<String>,
    pub election_officials: Option<Vec<ElectionOfficial>>,
    pub correspondence_address: Option<SimpleAddressType>,
    pub election_rules_url: Option<String>,
    pub voter_services: Option<Vec<String>>,
    pub ballot_info_url: Option<String>,
    pub hours_of_operation: Option<String>,
    pub physical_address: Option<SimpleAddressType>,
}

Information about an election administrative body (e.g. County Board of Elections).

This type is not used in any activity, and only used as part of another schema.

Fields

absentee_voting_info_url

A URL provided by this administrative body for information on absentee voting.

voting_location_finder_url

A URL provided by this administrative body for looking up where to vote.

name

The name of this election administrative body.

election_registration_confirmation_url

A URL provided by this administrative body for confirming that the voter is registered to vote.

election_info_url

A URL provided by this administrative body for looking up general election information.

election_registration_url

A URL provided by this administrative body for looking up how to register to vote.

election_officials

The election officials for this election administrative body.

correspondence_address

The mailing address of this administrative body.

election_rules_url

A URL provided by this administrative body describing election rules to the voter.

voter_services

A description of the services this administrative body may provide.

ballot_info_url

A URL provided by this administrative body to give contest information to the voter.

hours_of_operation

A description of the hours of operation for this administrative body.

physical_address

The physical address of this administrative body.