Struct google_civicinfo2::Contest

pub struct Contest {
    pub referendum_passage_threshold: Option<String>,
    pub referendum_text: Option<String>,
    pub office: Option<String>,
    pub referendum_brief: Option<String>,
    pub sources: Option<Vec<Source>>,
    pub primary_party: Option<String>,
    pub referendum_subtitle: Option<String>,
    pub electorate_specifications: Option<String>,
    pub roles: Option<Vec<String>>,
    pub number_elected: Option<String>,
    pub referendum_url: Option<String>,
    pub referendum_title: Option<String>,
    pub id: Option<String>,
    pub special: Option<String>,
    pub referendum_effect_of_abstain: Option<String>,
    pub number_voting_for: Option<String>,
    pub referendum_con_statement: Option<String>,
    pub district: Option<ElectoralDistrict>,
    pub level: Option<Vec<String>>,
    pub ballot_placement: Option<String>,
    pub referendum_ballot_responses: Option<Vec<String>>,
    pub referendum_pro_statement: Option<String>,
    pub candidates: Option<Vec<Candidate>>,
    pub type_: Option<String>,
}

Information about a contest that appears on a voter's ballot.

This type is not used in any activity, and only used as part of another schema.

Fields
referendum_passage_threshold: Option<String>

The threshold of votes that the referendum needs in order to pass, e.g. "two-thirds". This field is only populated for contests of type 'Referendum'.

referendum_text: Option<String>

The full text of the referendum. This field is only populated for contests of type 'Referendum'.

office: Option<String>

The name of the office for this contest.

referendum_brief: Option<String>

Specifies a short summary of the referendum that is typically on the ballot below the title but above the text. This field is only populated for contests of type 'Referendum'.

sources: Option<Vec<Source>>

A list of sources for this contest. If multiple sources are listed, the data has been aggregated from those sources.

primary_party: Option<String>

If this is a partisan election, the name of the party it is for.

referendum_subtitle: Option<String>

A brief description of the referendum. This field is only populated for contests of type 'Referendum'.

electorate_specifications: Option<String>

A description of any additional eligibility requirements for voting in this contest.

roles: Option<Vec<String>>

The roles which this office fulfills.

number_elected: Option<String>

The number of candidates that will be elected to office in this contest.

referendum_url: Option<String>

A link to the referendum. This field is only populated for contests of type 'Referendum'.

referendum_title: Option<String>

The title of the referendum (e.g. 'Proposition 42'). This field is only populated for contests of type 'Referendum'.

id: Option<String>

An ID for this object. IDs may change in future requests and should not be cached. Access to this field requires special access that can be requested from the Request more link on the Quotas page.

special: Option<String>

"Yes" or "No" depending on whether this a contest being held outside the normal election cycle.

referendum_effect_of_abstain: Option<String>

Specifies what effect abstaining (not voting) on the proposition will have (i.e. whether abstaining is considered a vote against it). This field is only populated for contests of type 'Referendum'.

number_voting_for: Option<String>

The number of candidates that a voter may vote for in this contest.

referendum_con_statement: Option<String>

A statement in opposition to the referendum. It does not necessarily appear on the ballot. This field is only populated for contests of type 'Referendum'.

district: Option<ElectoralDistrict>

Information about the electoral district that this contest is in.

level: Option<Vec<String>>

The levels of government of the office for this contest. There may be more than one in cases where a jurisdiction effectively acts at two different levels of government; for example, the mayor of the District of Columbia acts at "locality" level, but also effectively at both "administrative-area-2" and "administrative-area-1".

ballot_placement: Option<String>

A number specifying the position of this contest on the voter's ballot.

referendum_ballot_responses: Option<Vec<String>>

The set of ballot responses for the referendum. A ballot response represents a line on the ballot. Common examples might include "yes" or "no" for referenda. This field is only populated for contests of type 'Referendum'.

referendum_pro_statement: Option<String>

A statement in favor of the referendum. It does not necessarily appear on the ballot. This field is only populated for contests of type 'Referendum'.

candidates: Option<Vec<Candidate>>

The candidate choices for this contest.

type_: Option<String>

The type of contest. Usually this will be 'General', 'Primary', or 'Run-off' for contests with candidates. For referenda this will be 'Referendum'. For Retention contests this will typically be 'Retention'.