Struct google_civicinfo2::GeographicDivision

pub struct GeographicDivision {
    pub office_indices: Option<Vec<u32>>,
    pub also_known_as: Option<Vec<String>>,
    pub name: Option<String>,
}

Describes a political geography.

This type is not used in any activity, and only used as part of another schema.

Fields

office_indices

List of indices in the offices array, one for each office elected from this division. Will only be present if includeOffices was true (or absent) in the request.

also_known_as

Any other valid OCD IDs that refer to the same division.

Because OCD IDs are meant to be human-readable and at least somewhat predictable, there are occasionally several identifiers for a single division. These identifiers are defined to be equivalent to one another, and one is always indicated as the primary identifier. The primary identifier will be returned in ocd_id above, and any other equivalent valid identifiers will be returned in this list.

For example, if this division's OCD ID is ocd-division/country:us/district:dc, this will contain ocd-division/country:us/state:dc.

name

The name of the division.