Struct google_civicinfo2::RepresentativeInfoData

pub struct RepresentativeInfoData {
    pub divisions: Option<HashMap<String, GeographicDivision>>,
    pub officials: Option<Vec<Official>>,
    pub offices: Option<Vec<Office>>,
}

There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

representative info by division representatives (response)

Fields
divisions: Option<HashMap<String, GeographicDivision>>

Political geographic divisions that contain the requested address.

officials: Option<Vec<Official>>

Officials holding the offices listed above. Will only be present if includeOffices was true in the request.

offices: Option<Vec<Office>>

Elected offices referenced by the divisions listed above. Will only be present if includeOffices was true in the request.