Struct google_civicinfo2::RepresentativeInfoResponse

pub struct RepresentativeInfoResponse {
    pub divisions: Option<HashMap<String, GeographicDivision>>,
    pub kind: Option<String>,
    pub officials: Option<Vec<Official>>,
    pub offices: Option<Vec<Office>>,
    pub normalized_input: Option<SimpleAddressType>,
}

The result of a representative info lookup query.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

representative info by address representatives (response)

Fields

divisions

Political geographic divisions that contain the requested address.

kind

Identifies what kind of resource this is. Value: the fixed string "civicinfo#representativeInfoResponse".

officials

Officials holding the offices listed above. Will only be present if includeOffices was true in the request.

offices

Elected offices referenced by the divisions listed above. Will only be present if includeOffices was true in the request.

normalized_input

The normalized version of the requested address