Struct google_civicinfo2::VoterInfoResponse

pub struct VoterInfoResponse {
    pub early_vote_sites: Option<Vec<PollingLocation>>,
    pub drop_off_locations: Option<Vec<PollingLocation>>,
    pub other_elections: Option<Vec<Election>>,
    pub mail_only: Option<bool>,
    pub kind: Option<String>,
    pub normalized_input: Option<SimpleAddressType>,
    pub state: Option<Vec<AdministrationRegion>>,
    pub precinct_id: Option<String>,
    pub election: Option<Election>,
    pub polling_locations: Option<Vec<PollingLocation>>,
    pub contests: Option<Vec<Contest>>,
}

The result of a voter info lookup query.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

voter info query elections (response)

Fields

early_vote_sites

Locations where the voter is eligible to vote early, prior to election day.

drop_off_locations

Locations where a voter is eligible to drop off a completed ballot. The voter must have received and completed a ballot prior to arriving at the location. The location may not have ballots available on the premises. These locations could be open on or before election day as indicated in the pollingHours field.

other_elections

If no election ID was specified in the query, and there was more than one election with data for the given voter, this will contain information about the other elections that could apply.

mail_only

Specifies whether voters in the precinct vote only by mailing their ballots (with the possible option of dropping off their ballots as well).

kind

Identifies what kind of resource this is. Value: the fixed string "civicinfo#voterInfoResponse".

normalized_input

The normalized version of the requested address

state

Local Election Information for the state that the voter votes in. For the US, there will only be one element in this array.

precinct_id

no description provided

election

The election that was queried.

polling_locations

Locations where the voter is eligible to vote on election day.

contests

Contests that will appear on the voter's ballot.