Struct google_civicinfo2::Contest

pub struct Contest {
    pub office: Option<String>,
    pub sources: Option<Vec<Source>>,
    pub primary_party: Option<String>,
    pub electorate_specifications: Option<String>,
    pub number_elected: Option<String>,
    pub referendum_url: Option<String>,
    pub referendum_title: Option<String>,
    pub id: Option<String>,
    pub special: Option<String>,
    pub number_voting_for: Option<String>,
    pub district: Option<ElectoralDistrict>,
    pub roles: Option<Vec<String>>,
    pub level: Option<Vec<String>>,
    pub ballot_placement: Option<String>,
    pub referendum_subtitle: Option<String>,
    pub candidates: Option<Vec<Candidate>>,
    pub type_: Option<String>,
}

Information about a contest that appears on a voter's ballot.

This type is not used in any activity, and only used as part of another schema.

Fields

office

The name of the office for this contest.

sources

A list of sources for this contest. If multiple sources are listed, the data has been aggregated from those sources.

primary_party

If this is a partisan election, the name of the party it is for.

electorate_specifications

A description of any additional eligibility requirements for voting in this contest.

number_elected

The number of candidates that will be elected to office in this contest.

referendum_url

A link to the referendum. This field is only populated for contests of type 'Referendum'.

referendum_title

The title of the referendum (e.g. 'Proposition 42'). This field is only populated for contests of type 'Referendum'.

id

An ID for this object. IDs may change in future requests and should not be cached. Access to this field requires special access that can be requested from the Request more link on the Quotas page.

special

"Yes" or "No" depending on whether this a contest being held outside the normal election cycle.

number_voting_for

The number of candidates that a voter may vote for in this contest.

district

Information about the electoral district that this contest is in.

roles

The roles which this office fulfills.

level

The levels of government of the office for this contest. There may be more than one in cases where a jurisdiction effectively acts at two different levels of government; for example, the mayor of the District of Columbia acts at "locality" level, but also effectively at both "administrative-area-2" and "administrative-area-1".

ballot_placement

A number specifying the position of this contest on the voter's ballot.

referendum_subtitle

A brief description of the referendum. This field is only populated for contests of type 'Referendum'.

candidates

The candidate choices for this contest.

type_

The type of contest. Usually this will be 'General', 'Primary', or 'Run-off' for contests with candidates. For referenda this will be 'Referendum'.